[cgi] Fix bug: Handle query-string (e.g., for GET request method).
Regression introduced 2009-05-26, "Fix bug:
For (www cgi), don't mix query string and form data".
Reported by gregory benison.

This is a regression because we DO want to mix query-string
and form-data.  Previous to the 2009-05-26, the query-string
clobbered the form-data.  Still to figure out is whether or
not it is better to prefix or suffix the query-string pairs.

* source/cgi.scm (make-ccc): If ‘QUERY_STRING’ is non-empty,
parse it and prefix the resulting form variables to the others.